+++ docs/CHANGELOG.md
@@ -7,6 +7,14 @@ and this project orients towards [Semantic Versioning](http://semver.org/spec/v2
 Note: This project needs KSP to work and every new Ktorfit with an update of the KSP version is technically a breaking change.
 But there is no intent to bump the Ktorfit major version for every KSP update.
 
+# [TBD]()
+
+TBD
+========================================
+
+## Fixed
+- Fixed ksp generation for RequestType annotated Lists
+
 # [2.7.5]()
 
 2.7.5 - 2026-06-07