+++ docs/CHANGELOG.md
@@ -7,10 +7,23 @@ and this project orients towards [Semantic Versioning](http://semver.org/spec/v2
 Note: This project needs KSP to work and every new Ktorfit with an update of the KSP version is technically a breaking change.
 But there is no intent to bump the Ktorfit major version for every KSP update.
 
+* Supported KSP version: >=2.0.2
+* Supported Kotlin version: >=2.2.0
+* Ktor version: 3.4.1
+
+# [2.7.6]()
+
+2.7.6 - 2026-06-24
+* Supported KSP version: >=2.0.2
+* Supported Kotlin version: >=2.2.0
+* Ktor version: 3.5.0
+
+## Changed
+- Request builder now adds original path template as a request attribute.
+
 # [2.7.5]()
 
 2.7.5 - 2026-06-07
-========================================
 * Supported KSP version: >=2.0.2
 * Supported Kotlin version: >=2.2.0
 * Ktor version: 3.5.0
@@ -26,7 +39,6 @@ Thanks to @eygraber for contributing to this release!
 # [2.7.4]()
 
 2.7.4 - 2026-06-03
-========================================
 * Supported KSP version: >=2.0.2
 * Supported Kotlin version: >=2.2.0
 * Ktor version: 3.5.0